requirements.txt
streamlit
scikit-learn
joblib
numpy
pandas
shap 
matplotlib
fpdf2